requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
python-dotenv
httpx
sqlalchemy
qdrant-client
pdfplumber
psycopg2-binary
pgvector
google-generativeai
pydantic-settings
python-multipart
fpdf2
groq
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0

